+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release Notes
 
-## v2.8.2 (2024-08-10)
+## v2.8.3 (2024-08-10)
 - Fixed styles for the search field placeholder
 - Added 5 more recommended videos to the home page sidebar
 - Remove 5 outdated recommended videos from the home page sidebar